`initialize-instance :after job`; `macrolet await yield`
`job` now has an `initilialize-instance :after` method which registers
the job with its `job-queue` and the job it is `awaiting`, if
any. this allows `make-instance job` to be a public interface. job
also now has initargs for an `:executor`, which must be supplied, and
`:awaiting`, which defaults to `nil` but may be another `job`.

also, the `await` and `yield` macros now have global definitions which
error at macro-expand time, and are only bound to their actual
definitions within `async`.