`asynchronously` vs `with-executor` distinction; `job-seq`
this commit renames the macro previously called `with-executor` to
`asynchronously`, which i believe to be a better name for the macro
which evaluates its body as an asynchronous job in a new executor. it
also defines `with-executor` as a macro which evaluates its body
synchronously while `*job-queue*` is bound to a new executor.

it also defines `job-seq`, a function analogous to javascript's `Promise.then`.